a bunch of practice

+++ contests/nac/2024q/d.cpp
@@ -1,23 +1,23 @@
 #include <bits/stdc++.h>
 using namespace std;
 typedef long long ll;
-typedef pair<int, int> pii;
-map<int, vector<int>> tree;
-map<int, int> colours[100010];
-map<int, int> all_colours;
+typedef pair<ll, ll> pii;
+map<ll, vector<ll>> tree;
+map<ll, ll> colours[100010];
+map<ll, ll> all_colours;
 vector<pii> edges;
-int node_colors[100010];
-int ans[100010];
-int depth[100010];
-int sz[100010];
-int parent[100010];
-int n, u, v;
-void dfs(int node, int last, int r) {
+ll node_colors[100010];
+ll ans[100010];
+ll depth[100010];
+ll sz[100010];
+ll parent[100010];
+ll n, u, v;
+void dfs(ll node, ll last, ll r) {
     sz[node] = 1;
     depth[node] = r;
     parent[node] = node;
-    int max_parent = node;
-    for (int e: tree[node]) {
+    ll max_parent = node;
+    for (ll e: tree[node]) {
         if (e == last) continue;
         dfs(e, node, r + 1);
         if (sz[e] >= sz[max_parent]) {
@@ -27,7 +27,7 @@ void dfs(int node, int last, int r) {
     parent[node] = parent[max_parent];
     ans[node] = ans[max_parent];
     // printf("parent of %d is %d\n", node, parent[max_parent]);
-    for (int e: tree[node]) {
+    for (ll e: tree[node]) {
         if (e == last) continue;
         sz[node] += sz[e];
         // printf("n: %d | %d : %d\n", node, e, max_parent);
@@ -53,12 +53,12 @@ int main() {
     cin.tie(0);
     cin.sync_with_stdio(0);
     cin >> n;
-    for (int i = 1; i <= n; i++) {
+    for (ll i = 1; i <= n; i++) {
         cin >> node_colors[i];
         all_colours[node_colors[i]]++;
     }
 
-    for (int i = 1; i < n; i++) {
+    for (ll i = 1; i < n; i++) {
         cin >> u >> v;
         edges.push_back({u, v});
         tree[u].push_back(v);
@@ -71,7 +71,7 @@ int main() {
     // }
     for (auto [lower, upper]: edges) {
         if (depth[lower] < depth[upper]) swap(lower, upper);
-        printf("%d\n", ans[lower]);
+        printf("%lld\n", ans[lower]);
     }
 
     return 0;
